Move the code into a Python package
I separated the code into a file for the code that should be run locally and
a file for the code that should be run on Travis.
