+++ travis_docs_builder/travis.py
@@ -1,99 +1,12 @@
-#!/usr/bin/env python3
-import requests
-from requests.auth import HTTPBasicAuth
+"""
+The code that should be run on Travis
+"""
 
-from cryptography.hazmat.primitives.asymmetric import padding
-from cryptography.hazmat.backends import default_backend
-from cryptography.hazmat.primitives import serialization
-
-import json
-import base64
-from getpass import getpass
 import os
-import subprocess
-import shutil
-import sys
-import uuid
 import shlex
-
-def encrypt_variable(variable, repo, public_key=None):
-    """
-    Encrypt an environment variable for repo for Travis
-
-    ``variable`` should be a bytes object.
-
-    ``repo`` should be like 'gforsyth/travis_docs_builder'
-
-    ``public_key`` should be a pem format public key, obtained from Travis if
-    not provided.
-
-    """
-    if not isinstance(variable, bytes):
-        raise TypeError("variable should be bytes")
-
-    if not b"=" in variable:
-        raise ValueError("variable should be of the form 'VARIABLE=value'")
-
-    if not public_key:
-        # TODO: Error handling
-        r = requests.get('https://api.travis-ci.org/repos/{repo}/key'.format(repo=repo),
-            headers={'Accept': 'application/vnd.travis-ci.2+json'})
-        public_key = r.json()['key']
-
-    public_key = public_key.replace("RSA PUBLIC KEY", "PUBLIC KEY").encode('utf-8')
-    key = serialization.load_pem_public_key(public_key, backend=default_backend())
-
-    pad = padding.PKCS1v15()
-
-    return base64.b64encode(key.encrypt(variable, pad))
-
-class AuthenticationFailed(Exception):
-    pass
-
-def generate_GitHub_token(username, password=None, OTP=None, note=None, headers=None):
-    """
-    Generate a GitHub token for pushing from Travis
-
-    The scope requested is public_repo.
-
-    If no password or OTP are provided, they will be requested from the
-    command line.
-
-    The token created here can be revoked at
-    https://github.com/settings/tokens. The default note is
-    "travis_docs_builder token for pushing to gh-pages from Travis".
-    """
-    if not password:
-        password = getpass("Enter the GitHub password for {username}: ".format(username=username))
-
-    headers = headers or {}
-
-    if OTP:
-        headers['X-GitHub-OTP'] = OTP
-
-    auth = HTTPBasicAuth(username, password)
-    AUTH_URL = "https://api.github.com/authorizations"
-
-    note = note or "travis_docs_builder token for pushing to gh-pages from Travis"
-    data = {
-        "scopes": ["public_repo"],
-        "note": note,
-        "note_url": "https://github.com/gforsyth/travis_docs_builder",
-        "fingerprint": str(uuid.uuid4()),
-    }
-    r = requests.post(AUTH_URL, auth=auth, headers=headers, data=json.dumps(data))
-    if r.status_code == 401:
-        two_factor = r.headers.get('X-GitHub-OTP')
-        if two_factor:
-            print("A two-factor authentication code is required:", two_factor.split(';')[1].strip())
-            OTP = input("Authentication code: ")
-            return generate_GitHub_token(username=username, password=password,
-                OTP=OTP, note=note, headers=headers)
-
-        raise AuthenticationFailed("invalid username or password")
-
-    r.raise_for_status()
-    return r.json()['token']
+import shutil
+import subprocess
+import sys
 
 # XXX: Do this in a way that is streaming
 def run_command_hiding_token(args, token):
@@ -253,29 +166,3 @@ def push_docs():
         run(['git', 'push', '-q', 'origin_token', 'gh-pages'])
     else:
         print("The docs have not changed. Not updating")
-
-if __name__ == '__main__':
-    on_travis = os.environ.get("TRAVIS_JOB_NUMBER", '')
-
-    if on_travis:
-        # TODO: Get this automatically
-        repo = sys.argv[1]
-        if setup_GitHub_push(repo):
-            commit_docs()
-            push_docs()
-    else:
-        username = input("What is your GitHub username? ")
-        token = generate_GitHub_token(username)
-
-        repo = input("What repo to you want to build the docs for? ")
-        encrypted_variable = encrypt_variable("GH_TOKEN={token}".format(token=token).encode('utf-8'), repo=repo)
-        travis_content = """
-env:
-  global:
-    secure: "{encrypted_variable}"
-
-""".format(encrypted_variable=encrypted_variable.decode('utf-8'))
-
-        print("Put\n", travis_content, "in your .travis.yml.\n",
-            "Also make sure to create an empty gh-pages branch on GitHub, and "
-            "enable it at https://github.com/{repo}/settings".format(repo=repo), sep='')
